Don't save/restore the state when solution is found
When a solution is found, we want to keep the assignment,
so restoring it back is forgetting the solution.

Also, there seems to be a false-positive in cppcheck feature that suggests
to make methods const. cppcheck misses the case when the method needs to
pass non-const reference to internal data, which makes it impossible to
make the method cost (state_saver needs to get non-const domain
reference, so that it can restore the data back).